# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Cycle Shop Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface SaleInput {
  /** Sale date in YYYY-MM-DD format */
  date: string;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  cycleModel: string;
  cycleSize: string;
  cycleColor: string;
  supplier: string;
  inventoryId: number;
  invoiceNo: string;
  buyingPrice: number;
  /** Profit margin percentage */
  profitMargin: number;
  /** Discount percentage */
  discount: number;
  /** GST percentage */
  gst: number;
  modeOfPayment: string;
  /** @nullable */
  expenses?: number | null;
}

export interface Sale {
  id: number;
  serialNo: number;
  date: string;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  cycleModel: string;
  cycleSize: string;
  cycleColor: string;
  supplier: string;
  /** @nullable */
  inventoryId: number | null;
  invoiceNo: string;
  buyingPrice: number;
  profitMargin: number;
  sellingPrice: number;
  discount: number;
  gst: number;
  finalPrice: number;
  profitPercent: number;
  profitAmount: number;
  lossPercent: number;
  lossAmount: number;
  modeOfPayment: string;
  /** @nullable */
  expenses?: number | null;
  syncedToSheets?: boolean;
  createdAt: string;
}

export interface SaleListResponse {
  sales: Sale[];
  total: number;
  page: number;
  pageSize: number;
}

export interface InventoryInput {
  cycleModel: string;
  cycleSize: string;
  cycleColor: string;
  supplier: string;
  /** Buying price for this cycle variant */
  buyingPrice: number;
  /** Default profit margin percentage for this cycle variant */
  profitMargin: number;
  /** @minimum 0 */
  availableStock: number;
}

export interface Inventory {
  id: number;
  cycleModel: string;
  cycleSize: string;
  cycleColor: string;
  supplier: string;
  buyingPrice: number;
  profitMargin: number;
  availableStock: number;
  soldStock: number;
  balanceStock: number;
  createdAt: string;
}

export interface InventoryListResponse {
  inventory: Inventory[];
}

export interface SalesSummary {
  totalSales: number;
  totalRevenue: number;
  totalProfit: number;
  totalLoss: number;
  totalExpenses: number;
  averageMargin: number;
  profitableCount: number;
  lossCount: number;
  monthLabel: string;
  recentSales?: Sale[];
}

export interface ReportSummary {
  totalSales: number;
  totalRevenue: number;
  totalProfit: number;
  totalLoss: number;
  totalExpenses: number;
  netProfit: number;
}

export interface SalesReport {
  startDate: string;
  endDate: string;
  sales: Sale[];
  summary: ReportSummary;
}

export interface SyncResult {
  success: boolean;
  message: string;
  rowsSynced: number;
}

export type ListSalesParams = {
page?: number;
pageSize?: number;
/**
 * Filter by month (e.g. 2025-07)
 * @nullable
 */
month?: string | null;
};

export type GetSalesSummaryParams = {
/**
 * Month filter (e.g. 2025-07), defaults to current month
 * @nullable
 */
month?: string | null;
};

export type GetSalesReportParams = {
/**
 * Start date in YYYY-MM-DD format
 */
startDate: string;
/**
 * End date in YYYY-MM-DD format
 */
endDate: string;
};