# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Cycle Shop Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List all inventory entries
      responses:
        "200":
          description: List of inventory records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryListResponse"
    post:
      operationId: createInventory
      tags: [inventory]
      summary: Create a new inventory entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryInput"
      responses:
        "201":
          description: Inventory entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inventory"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Inventory item already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /inventory/{id}:
    get:
      operationId: getInventory
      tags: [inventory]
      summary: Get an inventory item by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory item found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inventory"
        "404":
          description: Inventory item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateInventory
      tags: [inventory]
      summary: Update an inventory item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryInput"
      responses:
        "200":
          description: Inventory item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inventory"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Inventory item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Inventory item already exists or cannot be updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteInventory
      tags: [inventory]
      summary: Delete an inventory item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory item deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid inventory ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Inventory item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Inventory item has linked sales
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    InventoryInput:
      type: object
      required:
        - cycleModel
        - cycleSize
        - cycleColor
        - supplier
        - buyingPrice
        - profitMargin
        - availableStock
      properties:
        cycleModel:
          type: string
        cycleSize:
          type: string
        cycleColor:
          type: string
        supplier:
          type: string
        buyingPrice:
          type: number
          description: "Buying price for this cycle variant"
        profitMargin:
          type: number
          description: "Default profit margin percentage for this cycle variant"
        availableStock:
          type: integer
          minimum: 0
    Inventory:
      type: object
      required:
        - id
        - cycleModel
        - cycleSize
        - cycleColor
        - supplier
        - buyingPrice
        - profitMargin
        - availableStock
        - soldStock
        - balanceStock
        - createdAt
      properties:
        id:
          type: integer
        cycleModel:
          type: string
        cycleSize:
          type: string
        cycleColor:
          type: string
        supplier:
          type: string
        buyingPrice:
          type: number
        profitMargin:
          type: number
        availableStock:
          type: integer
        soldStock:
          type: integer
        balanceStock:
          type: integer
        createdAt:
          type: string
    InventoryListResponse:
      type: object
      required:
        - inventory
      properties:
        inventory:
          type: array
          items:
            $ref: "#/components/schemas/Inventory"
...