# lib/api-zod/src/generated/types/inventory.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Cycle Shop Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Inventory {
  id: number;
  cycleModel: string;
  cycleSize: string;
  cycleColor: string;
  supplier: string;
  buyingPrice: number;
  profitMargin: number;
  availableStock: number;
  soldStock: number;
  balanceStock: number;
  createdAt: string;
}