# lib/api-zod/src/generated/types/inventoryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Cycle Shop Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryInput {
  cycleModel: string;
  cycleSize: string;
  cycleColor: string;
  supplier: string;
  /** Buying price for this cycle variant */
  buyingPrice: number;
  /** Default profit margin percentage for this cycle variant */
  profitMargin: number;
  /** @minimum 0 */
  availableStock: number;
}