# lib/api-zod/src/generated/types/inventoryListResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Cycle Shop Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { Inventory } from './inventory';

export interface InventoryListResponse {
  inventory: Inventory[];
}